+++ anvil/tests/it/main.rs
@@ -12,7 +12,8 @@ mod logs;
 mod proof;
 mod pubsub;
 // mod revert; // TODO uncomment <https://github.com/gakonst/ethers-rs/issues/2186>
-mod otterscan_api;
+#[cfg(feature = "otterscan")]
+mod otterscan;
 mod sign;
 mod traces;
 mod transaction;
